Added flag command parsing

+++ bin/prompt_chat.dart
@@ -369,13 +369,11 @@ void runApp(ChatAPI api) async {
               print("Please login to send a direct message.");
               break;
             }
-            print("Enter the message:");
-            var message = stdin.readLineSync();
-            if (message == null) {
-              print("Please enter a message.");
-              break;
-            }
-            api.sendDm(ccs[1], message, currUsername);
+            var recipient = getFlagValue("--recipient", currentCommand);
+            var message = getFlagValue("--message", currentCommand);
+            api.sendDm(recipient, message, currUsername);
+            print("Direct message sent successfully.");
+            break;
           }
         case "display-dms":
           {
@@ -383,19 +381,18 @@ void runApp(ChatAPI api) async {
               print("Please login to view direct messages.");
               break;
             }
-            List<String> messages = await api.getRecievedDms(currUsername);
-            for (var element in messages) {
-              print(element);
+            var filter = getFlagValue("--filter", currentCommand);
+            List<String> messages;
+            if (filter == "received") {
+              messages = await api.getRecievedDms(currUsername);
+            } else if (filter == "sent") {
+              messages = await api.getSentDms(currUsername);
+            } else {
+              messages = [
+                ...await api.getRecievedDms(currUsername),
+                ...await api.getSentDms(currUsername)
+              ];
             }
-            break;
-          }
-        case "display-sent-dms":
-          {
-            if (currUsername == null) {
-              print("Please login to view direct messages.");
-              break;
-            }
-            List<String> messages = await api.getSentDms(currUsername);
             for (var element in messages) {
               print(element);
             }
@@ -407,10 +404,17 @@ void runApp(ChatAPI api) async {
               print("Please login first.");
               break;
             }
-            await api.logoutUser(currUsername);
-            currUsername = null;
-            api.deleteUser(currUsername);
-            print("User deleted successfully.");
+            var confirmFlag = getFlagValue("--confirm", currentCommand);
+            if (confirmFlag.toLowerCase() == "yes" ||
+                confirmFlag.toLowerCase() == "y") {
+              await api.logoutUser(currUsername);
+              currUsername = null;
+              api.deleteUser(currUsername);
+              print("User deleted successfully.");
+            } else {
+              print("Please add --confirm yes to confirm user deletion.");
+            }
+            break;
           }
         default:
           {
